+++ CHANGELOG.md
@@ -32,6 +32,10 @@ are noticeable to end-users since the last release. For developers, this project
 
 ## [Unreleased]
 
+### Added
+
+- Add a new argument, `@popup-toggle --id <id>`, to set the ID of a popup directly ([#XX]).
+
 ## [0.4.0] - 2024-11-23
 
 ### Added
@@ -51,6 +55,7 @@ are noticeable to end-users since the last release. For developers, this project
 
 [#21]: https://github.com/loichyan/tmux-toggle-popup/pull/21
 [#23]: https://github.com/loichyan/tmux-toggle-popup/pull/23
+[#XX]: https://github.com/loichyan/tmux-toggle-popup/pull/XX
 
 ## [0.3.0] - 2024-10-21